third_party/blink/web_tests/http/tests/devtools/console/viewport-testing/console-key-links-expected.txt

Tests that console links are keyboard navigable.


Running: testNavigatingLinks
Evaluating: console.log("Before");console.log("Text around www.chromium.org/1a multiple links, www.chromium.org/1b");console.log("www.chromium.org/2");
Message count: 3
Setting focus in prompt:

Shift+Tab:
Viewport virtual selection: 2
activeElement: BUTTON.devtools-link.text-button.link-style
active text: console-key-links.js:31

ArrowUp:
Viewport virtual selection: 2
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/2

ArrowUp:
Viewport virtual selection: 2
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:31 www.chromium.org/2

ArrowUp:
Viewport virtual selection: 1
activeElement: BUTTON.devtools-link.text-button.link-style
active text: console-key-links.js:31

ArrowUp:
Viewport virtual selection: 1
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/1b

ArrowUp:
Viewport virtual selection: 1
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/1a

ArrowUp:
Viewport virtual selection: 1
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:31 Text around www.chromium.org/1a multiple links, www.chromium.org/1b

ArrowDown:
Viewport virtual selection: 1
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/1a

ArrowDown:
Viewport virtual selection: 1
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/1b

ArrowDown:
Viewport virtual selection: 1
activeElement: BUTTON.devtools-link.text-button.link-style
active text: console-key-links.js:31

ArrowDown:
Viewport virtual selection: 2
activeElement: DIV.console-message-wrapper.console-from-api.console-info-level.console-selected
active text: console-key-links.js:31 www.chromium.org/2

ArrowDown:
Viewport virtual selection: 2
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/2

ArrowDown:
Viewport virtual selection: 2
activeElement: BUTTON.devtools-link.text-button.link-style
active text: console-key-links.js:31

Running: testNavigatingLinksInStackTrace
Evaluating: fn1()
Message count: 1
Setting focus in prompt:

Shift+Tab:

ArrowUp:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/linkInErrMsg

ArrowDown:

ArrowUp:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/linkInErrMsg

ArrowUp:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:23 Custom error with link www.chromium.org/linkInErrMsg
fn1 @ foo.js:23
(anonymous) @ console-key-links.js:76

ArrowRight:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:23 Custom error with link www.chromium.org/linkInErrMsg
fn1 @ foo.js:23
(anonymous) @ console-key-links.js:76

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: BUTTON.devtools-link.text-button.link-style
active text: www.chromium.org/linkInErrMsg

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: BUTTON.devtools-link.text-button.link-style
active text: foo.js:23

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: BUTTON.devtools-link.text-button.link-style
active text: foo.js:23

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: BUTTON.devtools-link.text-button.link-style
active text: console-key-links.js:76

ArrowDown:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: BUTTON.devtools-link.text-button.link-style
active text: console-key-links.js:76

ArrowLeft:
Viewport virtual selection: 0
Is trace expanded: YES
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:23 Custom error with link www.chromium.org/linkInErrMsg
fn1 @ foo.js:23
(anonymous) @ console-key-links.js:76

ArrowLeft:
Viewport virtual selection: 0
Is trace expanded: NO
activeElement: DIV.console-message-wrapper.console-from-api.console-error-level.console-selected
active text: foo.js:23 Custom error with link www.chromium.org/linkInErrMsg
fn1 @ foo.js:23
(anonymous) @ console-key-links.js:76
